Allow to use a Value as a method decorator
This way it's possible to create dependent "sub-value" with the outer
computation depending just on the value's result. Let's make an example:

Suppose you have a calculated value v which is dependent on other reactive
values x y z, what happens if a reactive function uses v's value
normally? That the to be created computation (let's call it c)will be dependent on v's
dependecies, x y z.

A side effect of this will be that one of the three source changes, c
will receive an invalidation even if v's value remains unchanged. With
the new value's behavior v is decorated with the Value class and  then c
will depend only on v, with that in turn depend on x y z.

This way if v's value changes less frequently than x, y and z, the c
computation will be rerun less times.