Implement ``asend()`` from the consumers to the source
For this to have some sense, the action of a consumer sending a value into the
source context must be deterministic, especially in the case of async
operations. To have this in the context of the async iteration protocol means
that when ``asend()`` is used the tools used to create the communication
should allow for the iteration to begin, have a consume phase and then end as
soon as the source has nothing more to say. With this in mind, the next round
of communication should start another iteration from the beginning if it plans
to use ``asend()`` reliabily and so all the "suspension" code has been
dropped. Now the source gets consumed to the end anyway.